NewGraphite: change to deferred connect strategy
Defer creating a connection until later.  This results in not failing to
create a new graphite object when the connection cannot be made.

In our setup, we see services failing to startup because we cannot setup
g2g.  This is due to the graphite connection failing to be setup,
because the service is down, or a network issue of any kind.
Since each metric submission does a reconnect of the connection (thereby
dealing with the connection being broken somehow during the lifetime of
the graphite object) it is ok to skip the initial connection attempt
during object creation.  It in this case makes the code more resistant
against upstream graphite service failures.